+++ E-credit server/user/src/main/java/com/spring/user/Client/NotificationClient.java
@@ -10,5 +10,5 @@ import java.util.List;
 @FeignClient(name = "notification-service", url = "${application.config.notification-url}")
 public interface NotificationClient {
     @GetMapping("/user/{id}")
-    List<NotificationDTO> findAllNotificationByUser(@PathVariable("id") Integer id);
+    List<NotificationDTO> findAllNotificationByUser(@PathVariable("id") Long id);
 }
